+++ packages/stream_chat/CHANGELOG.md
@@ -1,8 +1,13 @@
-## To come
+## Upcoming
 
 🐞 Fixed
 
 - Fixed streamWatchers. Before it was always new, now it is possible to follow the watchers of a channel. 
+## Upcoming
+
+🔄 Changed
+
+- Cancelling a attachment upload now removes the attachment from the message.
 
 ## 5.3.0